Cleaned up code in package collection
